rename legacy files
