+++ cocos/assets/scripts.ts
